+++ Harbinton.API.Application/Dto/Common/BaseDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Harbinton.API.Application.Dto
+{
+    public class BaseDto
+    {
+        public int Amount { get; set; }
+        public string AccountNumberToTransfer { get; set; }
+    }
+}
