+++ Harbinton.API.Application/Dto/User/CreateUserDto.cs
@@ -1,0 +1,10 @@
+﻿namespace Harbinton.API.Application.Dto.User
+{
+    public class CreateUserDto
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Phone { get; set; }
+        public string Address { get; set; }
+    }
+}
