+++ Harbinton.API.Application/Dto/User/DisplayDetailsDto.cs
@@ -1,0 +1,11 @@
+﻿namespace Harbinton.API.Application.Dto.User
+{
+    public class DisplayDetailsDto
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Phone { get; set; }
+        public string Address { get; set; }
+        public AccountDto Account {get;set;}
+    }
+}
