+++ Harbinton.API.Application/Harbinton - Backup.API.Application.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Harbinton.API.Persitence\Harbinton.API.Persitence.csproj" />
+  </ItemGroup>
+
+</Project>
+Comm
