+++ Harbinton.API/Controllers/TransactionController.cs
@@ -1,6 +1,8 @@
-﻿using Harbinton.API.Contract;
-using Harbinton.API.Dto;
-using Harbinton.API.ResponseData;
+﻿
+using Harbinton.API.Application.Contracts.Persitence;
+using Harbinton.API.Application.Dto;
+using Harbinton.API.Application.Dto.Transaction;
+using Harbinton.API.Application.ResponseData;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
