--- Harbinton.API/Dto/BaseDto.cs
@@ -1,8 +1,0 @@
-﻿namespace Harbinton.API.Dto
-{
-    public class BaseDto
-    {
-        public int Amount { get; set; }
-        public string AccountNumberToTransfer { get; set; }
-    }
-}
