--- Harbinton.API/Dto/BillDto.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Harbinton.API.Dto
-{
-    public class BillDto : AccountDto
-    {
-        [Required]
-        public string BillType { get; set; }
-    }
-
-
-
-}
