--- Harbinton.API/Dto/CreateUserDto.cs
@@ -1,10 +1,0 @@
-﻿namespace Harbinton.API.Dto
-{
-    public class CreateUserDto
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Phone { get; set; }
-        public string Address { get; set; }
-    }
-}
